gro_recommendation_app/machine_learning.py
#python
# Import necessary libraries
from sklearn.neighbors import NearestNeighbors
import pandas as pd
import numpy as np

class Recommender:
    def __init__(self, data, n_neighbors=5):
        """
        This class is used to create a recommendation model.
        It takes a pandas DataFrame of data and the number of neighbors to consider for the recommendation.
        """
        self.data = data
        self.model = NearestNeighbors(n_neighbors=n_neighbors, algorithm='ball_tree')
        self.model.fit(self.data)

    def recommend(self, user_data):
        """
        This function uses the recommendation model to provide recommendations.
        It takes a pandas DataFrame of user data as input and returns a list of recommendations.
        """
        distances, indices = self.model.kneighbors(user_data)
        recommendations = self.data.iloc[indices[0]].index.tolist()
        return recommendations

def get_recommendations(user_data):
    """
    This function is used to get recommendations for a user.
    It takes a pandas DataFrame of user data as input and returns a list of recommendations.
    """
    # Load the data
    data = pd.read_csv('data.csv')

    # Initialize the recommender
    recommender = Recommender(data)

    # Get the recommendations
    recommendations = recommender.recommend(user_data)

    return recommendations

