gro_recommendation_app/test_machine_learning.py
#python
# Import necessary libraries
import pandas as pd
import numpy as np
from sklearn.neighbors import NearestNeighbors
import unittest
from unittest.mock import patch, MagicMock

# Import the function to test
from machine_learning import Recommender, get_recommendations

class TestRecommender(unittest.TestCase):
    def setUp(self):
        """
        This function sets up the necessary data for the tests.
        """
        self.data = pd.DataFrame({
            'age': [25, 30, 35, 40],
            'income': [50000, 60000, 70000, 80000],
            'product_id': [1, 2, 3, 4]
        })
        self.user_data = pd.DataFrame({
            'age': [30],
            'income': [60000],
            'product_id': [2]
        })
        self.recommender = Recommender(self.data)

    def test_init(self):
        """
        This function tests the __init__ function of the Recommender class.
        """
        self.assertEqual(self.recommender.data.equals(self.data), True)
        self.assertIsInstance(self.recommender.model, NearestNeighbors)

    def test_recommend(self):
        """
        This function tests the recommend function of the Recommender class.
        """
        recommendations = self.recommender.recommend(self.user_data)
        self.assertIsInstance(recommendations, list)
        self.assertEqual(len(recommendations), 5)

@patch('pandas.read_csv')
@patch('machine_learning.Recommender')
def test_get_recommendations(mock_Recommender, mock_read_csv):
    """
    This function tests the get_recommendations function.
    """
    # Set up the mock objects
    mock_read_csv.return_value = pd.DataFrame()
    mock_recommender = MagicMock()
    mock_Recommender.return_value = mock_recommender
    mock_recommender.recommend.return_value = [1, 2, 3, 4, 5]

    # Call the function
    recommendations = get_recommendations(self.user_data)

    # Check the results
    mock_read_csv.assert_called_once_with('data.csv')
    mock_Recommender.assert_called_once_with(mock_read_csv.return_value)
    mock_recommender.recommend.assert_called_once_with(self.user_data)
    self.assertEqual(recommendations, [1, 2, 3, 4, 5])

if __name__ == '__main__':
    unittest.main()

